+++ GameServerScripts/quests/Midgard/StolenEggs.cs
@@ -549,7 +549,7 @@ namespace DOL.GS.Quests.Midgard
 			grifflet.Model = 1236;
 			grifflet.Name = "Grifflet";
 			grifflet.GuildName = "Part of " + m_questPlayer.GetName(0, false) + "'s " + questTitle + " Quest";
-			grifflet.Flags ^= (uint)GameNPC.eFlags.PEACE;
+			grifflet.Flags ^= GameNPC.eFlags.PEACE;
 			grifflet.CurrentRegionID = askefruerTrainer.CurrentRegionID;
 			grifflet.Size = 20;
 			grifflet.Level = 3;
